pyproject.toml
[tool.poetry]
name = "aws-codepipeline-discord-notifications"
version = "0.1.0"
description = ""
authors = ["banananeer <banananeer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aws-lambda-powertools = "^2.26.0"
discord-webhook = "^1.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
aws-lambda-powertools = {extras = ["aws-sdk"], version = "^2.26.0"}
pytest-responses = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

